backend/requirements.txt
Django==3.1.7
django-fsm==2.7.1
django-fsm-admin==1.2.4
django-fsm-log==2.0.1
